+++ packages/dash-table/demo/App.js
@@ -23,9 +23,7 @@ class App extends Component {
                 dataframe: dataframe,
                 columns: clone(mockData.columns).map(col => R.merge(col, {
                     editable_name: true,
-                    deletable: true,
-                    minWidth: col.width,
-                    maxWidth: col.width
+                    deletable: true
                 //     type: 'dropdown'
                 })),
                 editable: true,
@@ -44,9 +42,16 @@ class App extends Component {
                         }))
                     }
                 ],
-                table_style: [
-                    { selector: '.dash-spreadsheet.dash-freeze-left', rule: 'width: 1000px; max-width: 1000px;' }
-                ]
+                style_table: {
+                    max_width: '1000px',
+                    width: '1000px'
+                },
+                style_cells: [
+                    { max_width: 150, min_width: 150, width: 150 },
+                    { if: { column_id: 'rows' }, max_width: 40, min_width: 40, width: 40 },
+                    { if: { column_id: 'bbb' }, max_width: 200, min_width: 200, width: 200 },
+                    { if: { column_id: 'bbb-readonly' }, max_width: 200, min_width: 200, width: 200 }
+                ],
             }
         };