+++ packages/dash-table/src/dash-table/components/CellInput/types.ts
@@ -1,5 +1,3 @@
-import { CSSProperties } from 'react';
-
 export interface IDropdownOption {
     label: string;
     value: string;
@@ -10,13 +8,4 @@ export type IDropdownOptions = IDropdownOption[];
 export interface IConditionalDropdown {
     condition: string;
     dropdown: IDropdownOptions;
-}
-
-export interface IStyle {
-    target?: undefined;
-    style: CSSProperties;
-}
-
-export interface IConditionalStyle extends IStyle {
-    condition: string;
 }