+++ packages/dash-table/src/dash-table/components/HeaderFactory.tsx
@@ -2,7 +2,7 @@ import * as R from 'ramda';
 import React from 'react';
 
 import { arrayMap } from 'core/math/arrayZipMap';
-import { matrixMap } from 'core/math/matrixZipMap';
+import { matrixMap3 } from 'core/math/matrixZipMap';
 
 import { ControlledTableProps } from 'dash-table/components/Table/props';
 import derivedHeaderContent from 'dash-table/derived/header/content';
@@ -11,6 +11,8 @@ import getIndices from 'dash-table/derived/header/indices';
 import getLabels from 'dash-table/derived/header/labels';
 import derivedHeaderOperations from 'dash-table/derived/header/operations';
 import derivedHeaderWrappers from 'dash-table/derived/header/wrappers';
+import { derivedRelevantHeaderStyles } from 'dash-table/derived/style';
+import derivedHeaderStyles from 'dash-table/derived/header/wrapperStyles';
 
 export default class HeaderFactory {
     private readonly headerContent = derivedHeaderContent();
@@ -21,9 +23,11 @@ export default class HeaderFactory {
         return this.propsFn();
     }
 
-    constructor(private readonly propsFn: () => ControlledTableProps) {
-
-    }
+    constructor(
+        private readonly propsFn: () => ControlledTableProps,
+        private readonly headerStyles = derivedHeaderStyles(),
+        private readonly relevantStyles = derivedRelevantHeaderStyles()
+    ) { }
 
     public createHeaders() {
         const props = this.props;
@@ -37,7 +41,9 @@ export default class HeaderFactory {
             setProps,
             sorting,
             sorting_settings,
-            sorting_type
+            sorting_type,
+            style_cells_and_headers,
+            style_headers
         } = props;
 
         const headerRows = getHeaderRows(columns);
@@ -51,6 +57,14 @@ export default class HeaderFactory {
             headerRows,
             row_selectable,
             row_deletable
+        );
+
+        const relevantStyles = this.relevantStyles(style_cells_and_headers, style_headers);
+
+        const wrapperStyles = this.headerStyles(
+            columns,
+            headerRows,
+            relevantStyles
         );
 
         const wrappers = this.headerWrappers(
@@ -70,7 +84,11 @@ export default class HeaderFactory {
             props
         );
 
-        const headers = matrixMap(wrappers, content, (w, c) => React.cloneElement(w, { children: [c] }));
+        const headers = matrixMap3(
+            wrappers,
+            wrapperStyles,
+            content,
+            (w, s, c) => React.cloneElement(w, { children: [c], style: s }));
 
         return arrayMap(operations, headers, (o, h) => Array.prototype.concat(o, h));
     }