+++ packages/dash-table/tests/dash/app_sizing.py
@@ -295,15 +295,16 @@ def layout():
                     dataframe=df.to_dict("rows"),
                     content_style="grow",
                     columns=[
-                        {"name": i, "id": i, "width": "16.67%"} for i in df.columns
+                        {"name": i, "id": i} for i in df.columns
                     ],
-                    table_style=[
+                    css=[
                         {"selector": ".dash-spreadsheet", "rule": "width: 100%"},
                         {
                             "selector": ".dash-cell[data-dash-column=Region]",
                             "rule": "white-space: normal",
                         },
                     ],
+                    style_cells=[{"width": "16.67%"}]
                 ),
                 section_title("Dash Table - Single Column Width by Percent"),
                 html.Div(
@@ -321,16 +322,17 @@ def layout():
                     dataframe=df.to_dict("rows"),
                     content_style="grow",
                     columns=[
-                        {"name": i, "id": i, "width": "50%" if i == "Region" else None}
+                        {"name": i, "id": i }
                         for i in df.columns
                     ],
-                    table_style=[
+                    css=[
                         {"selector": ".dash-spreadsheet", "rule": "width: 100%"},
                         {
                             "selector": ".dash-cell[data-dash-column=Region]",
                             "rule": "white-space: normal",
                         },
                     ],
+                    style_cells=[{ "if": { "column_id": "Region" }, "width": "50%" }]
                 ),
                 section_title("Dash Table - Underspecified Widths"),
                 html.Div(
@@ -347,18 +349,14 @@ def layout():
                         {
                             "name": i,
                             "id": i,
-                            "width": "100px"
-                            if i == "Dem" or i == "Rep" or i == "Ind"
-                            else None,
-                            "minWidth": "100px"
-                            if i == "Dem" or i == "Rep" or i == "Ind"
-                            else None,
-                            "maxWidth": "100px"
-                            if i == "Dem" or i == "Rep" or i == "Ind"
-                            else None,
                         }
                         for i in df.columns
                     ],
+                    style_cells=[
+                        { "if": { "column_id": "Dem" }, "width": "100px", "min_width": "100px", "max_width": "100px" },
+                        { "if": { "column_id": "Rep" }, "width": "100px", "min_width": "100px", "max_width": "100px" },
+                        { "if": { "column_id": "Ind" }, "width": "100px", "min_width": "100px", "max_width": "100px" }
+                    ]
                 ),
                 section_title("Dash Table - Widths that are smaller than the content"),
                 html.Div(
@@ -372,14 +370,15 @@ def layout():
                     id="sizing-5",
                     dataframe=df.to_dict("rows"),
                     columns=[
-                        {"name": i, "id": i, "width": "100px"} for i in df.columns
+                        {"name": i, "id": i } for i in df.columns
                     ],
-                    table_style=[
+                    css=[
                         {
                             "selector": ".dash-cell[data-dash-column=Region]",
                             "rule": "white-space: normal",
                         }
                     ],
+                    style_cells=[{ "width": "100px" }]
                 ),
                 section_title(
                     "Dash Table - Widths that are smaller than the content (forced)"
@@ -398,18 +397,18 @@ def layout():
                         {
                             "name": i,
                             "id": i,
-                            "width": "100px",
-                            "minWidth": "100px",
-                            "maxWidth": "100px",
                         }
                         for i in df.columns
                     ],
-                    table_style=[
+                    css=[
                         {
                             "selector": ".dash-cell[data-dash-column=Region]",
                             "rule": "white-space: normal",
                         }
                     ],
+                    style_cells=[
+                        { "width": "100px", "min_width": "100px", "max_width": "100px" }
+                    ]
                 ),
                 section_title("Dash Table - Content with Ellipses"),
                 # ...