+++ packages/dash-table/tests/dash/app_styling.py
@@ -352,7 +352,14 @@ def layout():
             # ...
 
             section_title('Dash Table - Row Padding'),
-            # ...
+            dash_table.Table(
+                id="styling-2",
+                dataframe=df.to_dict("rows"),
+                columns=[
+                    {"name": i, "id": i} for i in df.columns
+                ],
+                style_cells=[{ "padding_bottom": 5, "padding_top": 5}]
+            ),
 
             section_title('Dash Table - List Style with Minimal Headers'),
             # ...
@@ -364,7 +371,37 @@ def layout():
             # ...
 
             section_title('Dash Table - Dark Theme with Cells'),
-            # ...
+            dash_table.Table(
+                id="styling-6",
+                dataframe=df.to_dict("rows"),
+                columns=[
+                    {"name": i, "id": i} for i in df.columns
+                ],
+                content_style="grow",
+                style_table={
+                    "width": "100%"
+                },
+                style_cells=[{
+                    "background_color": "rgb(50, 50, 50)",
+                    "color": "white",
+                    "font_family": "arial"
+                }, {
+                    "if": { "column_id": "Humidity" },
+                    "font_family": "monospace",
+                    "padding_left": 20,
+                    "text_align": "left"
+                }, {
+                    "if": { "column_id": "Pressure" },
+                    "font_family": "monospace",
+                    "padding_left": 20,
+                    "text_align": "left"
+                }, {
+                    "if": { "column_id": "Temperature" },
+                    "font_family": "monospace",
+                    "padding_left": 20,
+                    "text_align": "left"
+                }]
+            ),
 
             section_title('Dash Table - Dark Theme with Rows'),
             # ...
@@ -373,10 +410,45 @@ def layout():
             # ...
 
             section_title('Dash Table - Highlighting Certain Columns'),
-            # ...
+            dash_table.Table(
+                id="styling-9",
+                dataframe=df.to_dict("rows"),
+                columns=[
+                    {"name": i, "id": i} for i in df.columns
+                ],
+                content_style="grow",
+                style_table={
+                    "width": "100%"
+                },
+                style_cells=[{
+                    "color": "rgb(60, 60, 60)",
+                    "padding_left": 20,
+                    "text-align": "left",
+                    "width": "20%"
+                }, {
+                    "if": { "column_id": "Temperature" },
+                    "background_color": "yellow"
+                }]
+            ),
 
             section_title('Dash Table - Highlighting Certain Cells'),
-            # ...
-
+            dash_table.Table(
+                id="styling-10",
+                dataframe=df.to_dict("rows"),
+                columns=[
+                    {"name": i, "id": i} for i in df.columns
+                ],
+                content_style="grow",
+                style_table={
+                    "width": "100%"
+                },
+                style_cells=[{
+                    "if": { "column_id": "Region", "filter": "Region eq str(Montreal)" },
+                    "background_color": "yellow"
+                }, {
+                    "if": { "column_id": "Humidity", "filter": "Humidity eq num(20)" },
+                    "background_color": "yellow"
+                }]
+            )
         ],
     )